Break dashboard into multiple files
